Change int to long for ids

+++ homework-06/src/main/java/ru/otus/homework06/dao/AuthorDao.java
@@ -10,7 +10,7 @@ public interface AuthorDao {
 
     int count();
 
-    Optional<Author> getById(int id);
+    Optional<Author> getById(long id);
 
     Author getBySurname(String surname);
 
@@ -22,6 +22,6 @@ public interface AuthorDao {
 
     void insert(Author author);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
 }
